[U] Damage item now.
