+++ .gitignore
@@ -3,3 +3,4 @@
 .gradle
 
 *.jar
+.idea/
