+++ build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'java'
+    id 'com.rikonardo.papermake' version '1.0.6'
 }
 
 group = 'org.dispensermining'
