+++ gradle.properties
@@ -1,0 +1,3 @@
+pmake.version=1.16.5
+pmake.args=-o=false
+pmake.autoop=true
