+++ src/main/resources/plugin.yml
@@ -1,4 +1,4 @@
 name: Dispenser-Mining-Paper
 version: '${version}'
-main: org.dispensermining.dispenserminingpaper.Dispenser_Mining_Paper
+main: io.github.elihuso.dispenseminingpaper.Dispenser_Mining_Paper
 api-version: '1.16'
